requirements.txt
cachetools==5.3.2
click==8.1.7
google-auth==2.23.3
google-auth-oauthlib==1.1.0
gspread==5.12.0
gspread-pandas==3.2.3
httplib2==0.22.0
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.2.2
pandas==2.0.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
requests-oauthlib==1.3.1
rsa==4.9
tzdata==2023.3
